update condor_status & glidein_resource ES scripts to ES 8.14
